+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/db/derived/DbDerivedDataRepository.kt
@@ -6,6 +6,7 @@ import cash.z.ecc.android.sdk.internal.repository.DerivedDataRepository
 import cash.z.ecc.android.sdk.model.AccountUuid
 import cash.z.ecc.android.sdk.model.BlockHeight
 import cash.z.ecc.android.sdk.model.FirstClassByteArray
+import cash.z.ecc.android.sdk.model.TransactionId
 import cash.z.ecc.android.sdk.model.TransactionRecipient
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -52,15 +53,22 @@ internal class DbDerivedDataRepository(
     override val allTransactions: Flow<List<DbTransactionOverview>>
         get() = invalidatingFlow.map { derivedDataDb.allTransactionView.getAllTransactions().toList() }
 
-    override fun getOutputProperties(transactionId: FirstClassByteArray) =
+    override fun getOutputProperties(transactionId: TransactionId) =
         derivedDataDb.txOutputsView
-            .getOutputProperties(transactionId)
+            .getOutputProperties(transactionId.value)
 
-    override fun getTransactionsByMemoSubstring(query: String): Flow<List<FirstClassByteArray>> =
-        derivedDataDb.txOutputsView.getTransactionsByMemoSubstring(query)
+    override fun getTransactionsByMemoSubstring(query: String): Flow<List<TransactionId>> =
+        derivedDataDb
+            .txOutputsView
+            .getTransactionsByMemoSubstring(query)
+            .map { listOfFirstClassByteArray ->
+                listOfFirstClassByteArray.map {
+                    TransactionId(it)
+                }
+            }
 
-    override fun getRecipients(transactionId: FirstClassByteArray): Flow<TransactionRecipient> {
-        return derivedDataDb.txOutputsView.getRecipients(transactionId)
+    override fun getRecipients(transactionId: TransactionId): Flow<TransactionRecipient> {
+        return derivedDataDb.txOutputsView.getRecipients(transactionId.value)
     }
 
     override suspend fun close() {